+++ index.js
@@ -8,7 +8,7 @@ port = process.env.PORT || 3000;
 
 mongoose = require('mongoose')
 const uri = process.env.mongodb;
-require('./Utils/api/models/models') //created model loading here
+require('./api/models/models') //created model loading here
 // mongoose instance connection url connection
 mongoose.connect(uri, {
  useNewUrlParser: true,
